Shotgun/Dash
To dash you need to be in the air and hit 'E'. Had to turn of kinematic on the rigid body to put it in, so if stuff breaks that's where the issue is.

Also includes the shotgun spawner and weapon swapping as well as UI update to show your active weapon. One you picked up the shotgun, hit 'Q' to swap between weapons.